Use `tail --pid` instead of a `sleep` loop
The steam client keeps a journal of all the processes spawned from launching a game (presumably to be able to kill all processes associated with a game). I've noticed, that every time sleep is called, a new "sleep" process spawns, which is logged by the steam client. However, when the sleep process exits, steam does not seem to remove it from its journal. Considering one sleep process is spawned per second, long play sessions are likely to spam the journal with exorbitant amounts of data.

This commit uses `tail --pid` to wait for the game to exit. This is functionally equivalent to the sleep loop, but does not create 1 process per second.